fix: start tryna fix event subscriptions
