+++ package.json
@@ -13,11 +13,13 @@
   },
   "dependencies": {
     "@radix-ui/react-dialog": "1.0.2",
+    "@soroban-react/chains": "^5.0.3",
+    "@soroban-react/events": "^5.0.3",
     "@stellar/freighter-api": "^1.4.0",
+    "abundance-token": "file:target/js-clients/abundance-token",
     "axios": "^0.27.2",
     "bigint-conversion": "^2.4.1",
     "crowdfund-contract": "file:target/js-clients/crowdfund-contract",
-    "abundance-token": "file:target/js-clients/abundance-token",
     "humanize-duration": "^3.27.3",
     "moment": "^2.29.4",
     "next": "^13.4.4",
